refactor(grid): refactor blender grid recipe, add further responsive classes
