+++ components/grid/blender/bootstrap/bootstrap-grid.js
@@ -1,0 +1,7 @@
+import { containers } from './containers';
+import { grid } from './grid';
+
+export const bootstrapGrid = (breakpoints) => ({
+	...containers,
+	...grid(breakpoints),
+});
