+++ components/grid/blender/bootstrap/mixins/container.js
@@ -1,0 +1,10 @@
+// Container mixins
+
+export const makeContainer = (paddingX) => ({
+	boxSizing: 'border-box',
+	width: '100%',
+	paddingLeft: paddingX,
+	paddingRight: paddingX,
+	marginLeft: 'auto',
+	marginRight: 'auto',
+});
